+++ spec/avro/dsl/test/key_with_optional.rb
@@ -1,0 +1,4 @@
+record :key_with_optional, namespace: :test do
+  required :id, :long
+  optional :name, :string
+end